feature: Implement lazy routes to optimize loading and reset ShoppingCart upon user logout.
