+++ src/pages/AboutPage.tsx
@@ -1,4 +1,3 @@
-import Footer from "../components/Footer";
 import Clients from "../components/Clients";
 import MemberCard from "../components/MemberCard";
 import { Icon } from "@iconify/react";
@@ -7,12 +6,10 @@ import happyWoman from "../assets/about/none.svg";
 import videoPreview from "../assets/about/videoPreview.jpeg";
 import member1 from "../assets/about/member1.jpeg";
 import workWithUs from "../assets/about/workWithUs.jpeg";
-import Header from "../components/Header";
 
 const AboutPage = () => {
   return (
     <>
-      <Header />
       <section
         id="hero"
         className="flex flex-wrap justify-center px-[11%] pt-10"
@@ -146,8 +143,6 @@ const AboutPage = () => {
           />
         </div>
       </section>
-
-      <Footer />
     </>
   );
 };
