+++ src/pages/TeamPage.tsx
@@ -1,18 +1,14 @@
 import { Icon } from "@iconify/react";
-import Footer from "../components/Footer";
 import MemberCard from "../components/MemberCard";
 import { teamMembers } from "../data";
-
 import hero1 from "../assets/members/hero1.jpeg";
 import hero2 from "../assets/members/hero2.jpeg";
 import hero3 from "../assets/members/hero3.jpeg";
 import hero4 from "../assets/members/hero1.jpeg";
-import Header from "../components/Header";
 
 const TeamPage = () => {
   return (
     <>
-      <Header />
       <section id="hero">
         <div className="flex flex-col items-center gap-5 py-20  ">
           <p className="text-center text-neutral-500 text-base font-bold font-['Montserrat'] tracking-tight ">
@@ -72,7 +68,6 @@ const TeamPage = () => {
           <Icon icon="devicon:linkedin" className="w-6 h-6" />
         </div>
       </section>
-      <Footer />
     </>
   );
 };
